+++ config/settings/local.py
@@ -20,10 +20,22 @@ ALLOWED_HOSTS = [
 # https://docs.djangoproject.com/en/dev/ref/settings/#caches
 CACHES = {
     'default': {
-        'BACKEND': 'django.core.cache.backends.locmem.LocMemCache',
-        'LOCATION': ''
-    }
+        'BACKEND': env.str("DJANGO_DEFAULT_CACHE_BACKEND",
+                           default='django_redis.cache.RedisCache'),
+        'LOCATION': env.str("DJANGO_DEFAULT_CACHE_LOCATION", default='redis://localhost:6379/0'),
+        'OPTIONS': {
+            'CLIENT_CLASS': "django_redis.client.DefaultClient",
+        }
+    },
+    # "select2": {
+    #     "BACKEND": "django_redis.cache.RedisCache",
+    #     "LOCATION": env.str("DJANGO_SELECT2_CACHE_LOCATION", default='redis://127.0.0.1:6379/2'),
+    #     "OPTIONS": {
+    #         "CLIENT_CLASS": "django_redis.client.DefaultClient",
+    #     }
+    # }
 }
+SELECT2_CACHE_BACKEND = "default"
 
 # TEMPLATES
 # ------------------------------------------------------------------------------
@@ -39,8 +51,8 @@ EMAIL_PORT = 1025
 
 # STATIC FILES
 # ------------------------------------------------------------------------------
-STATICFILES_STORAGE = 'whitenoise.storage.CompressedManifestStaticFilesStorage'
-#STATICFILES_STORAGE = 'whitenoise.storage.ManifestStaticFilesStorage'
+#STATICFILES_STORAGE = 'whitenoise.storage.CompressedManifestStaticFilesStorage'
+STATICFILES_STORAGE = 'whitenoise.storage.ManifestStaticFilesStorage'
 # with 'whitenoise.storage.CompressedManifestStaticFilesStorage', collect static is
 # very slow in development, not sure why - with production settings, it's fast
 
@@ -148,7 +160,7 @@ LOGGING = {
             'propagate': True
         },
         'topobank.taskapp.tasks': {
-            'level': 'INFO',
+            'level': 'DEBUG',
             'handlers': ['console'],
             'propagate': True
         },