+++ topobank/manager/containers.py
@@ -5,6 +5,7 @@ import zipfile
 import os.path
 import yaml
 import textwrap
+import json
 import logging
 
 from django.utils.timezone import now
@@ -16,7 +17,7 @@ from .models import Topography
 _log = logging.getLogger(__name__)
 
 
-def write_surface_container(file, surfaces, request=None):
+def write_surface_container(file, surfaces):
     """Write container data to a file.
 
     Parameters
@@ -25,8 +26,6 @@ def write_surface_container(file, surfaces, request=None):
         Should be opened in "w" mode.
     surfaces: sequence of Surface instances
         Surface which should be included in container.
-    request: HTTPRequest
-        If None, urls of published surfaces will only be relative.
 
     Returns
     -------
@@ -103,7 +102,7 @@ def write_surface_container(file, surfaces, request=None):
 
             topography_dicts.append(topo_dict)
 
-        surface_dict = surface.to_dict(request)
+        surface_dict = surface.to_dict()
         surface_dict['topographies'] = topography_dicts
 
         surfaces_dicts.append(surface_dict)
@@ -151,7 +150,16 @@ def write_surface_container(file, surfaces, request=None):
                settings.TOPOBANK_VERSION))
 
     if len(publications) > 0:
+        #
+        # Add datacite_json
+        #
+        for pub in publications:
+            if pub.doi_name:
+                zf.writestr(f"other/datacite-{pub.short_url}.json", json.dumps(pub.datacite_json))
 
+        #
+        # Add license information to README
+        #
         licenses_used = set(pub.license for pub in publications)
         readme_txt += textwrap.dedent("""
         License information