+++ topobank/manager/forms.py
@@ -1,7 +1,10 @@
+import re
+
 from django.forms import forms, ModelMultipleChoiceField
 from django import forms
 from django_select2.forms import ModelSelect2MultipleWidget
 from django.contrib.postgres.forms import JSONField as JSONField4Form
+from django.conf import settings
 import bleach  # using bleach instead of django.utils.html.escape because it allows more (e.g. for markdown)
 
 from crispy_forms.helper import FormHelper
@@ -732,8 +735,7 @@ class SurfacePublishForm(forms.Form):
                                         help_text="""Please make sure you're not publishing data """
                                                   """from others without their authorization.""")
 
-    authors = forms.CharField(max_length=MAX_LEN_AUTHORS_FIELD, required=False)  # we be filled in clean() method
-    num_author_fields = forms.IntegerField(required=True)
+    authors_json = forms.JSONField(required=True)
 
     helper = FormHelper()
     helper.form_method = 'POST'
@@ -744,7 +746,7 @@ class SurfacePublishForm(forms.Form):
     helper.layout = Layout(
         Div(
             HTML('<h2 class="alert-heading">Please enter the authors</h2>'),
-            Field('authors', template="manager/multi_author_field.html"),
+            Field('authors_json', template="manager/multi_author_field.html"),
             css_class="alert alert-primary"
         ),
         Div(
@@ -766,39 +768,91 @@ class SurfacePublishForm(forms.Form):
             css_class="alert alert-primary"),
     )
 
-    def __init__(self, *args, **kwargs):
-        num_author_fields = kwargs.pop('num_author_fields', 1)
-        super().__init__(*args, **kwargs)
-
-        for i in range(num_author_fields):
-            self.fields[f'author_{i}'] = forms.CharField(required=False, label=f"{i + 1}. Author")
-
     def clean(self):
         cleaned_data = super().clean()
 
-        authors = []
+        #
+        # Check each author
+        #
+        # - Check that only valid fields are included
+        # - bleach all fields in order to prevent XSS attacks
+        # - strip text fields
+        # - first and last name given?
+        # - if ORCID given, correct format?
+        # - if affiliation given, is there a name? If no data given, remove.
+        # - if affiliation ROR ID given, correct format?
+        # - has this author already given with same data?
 
-        for i in range(self.cleaned_data.get('num_author_fields')):
-            field_name = f'author_{i}'
-            author = self.cleaned_data.get(field_name)
-            if author:
-                author = author.strip()
-                if author in authors:
-                    raise forms.ValidationError("Author '%(author)s' is already in the list.",
-                                                code='duplicate_author',
-                                                params={'author': author})
-                elif len(author) > 0:
-                    authors.append(author)
+        authors = self.cleaned_data.get('authors_json')
 
-        if len(authors) == 0:
+        if (authors is None) or (len(authors) == 0):
             raise forms.ValidationError("At least one author must be given.")
 
-        authors_string = ", ".join(authors)
-        if len(authors_string) > MAX_LEN_AUTHORS_FIELD:
-            msg = """Representation of authors is too long, at maximum %(max_len)s characters are allowed."""
-            raise forms.ValidationError(msg, code='authors_too_long',
-                                        params=dict(max_len=MAX_LEN_AUTHORS_FIELD))
+        if len(authors) > settings.PUBLICATION_MAX_NUM_AUTHORS:
+            raise forms.ValidationError(f"Too many authors given, at maximum {settings.PUBLICATION_MAX_NUM_AUTHORS}"
+                                        "allowed.")
+        try:
+            for a in authors:
+                for k in a.keys():
+                    if k not in ['first_name', 'last_name', 'orcid_id', 'affiliations']:
+                        raise forms.ValidationError(f"Invalid key {k} given in author definition.")
+                for k in ['first_name', 'last_name', 'orcid_id']:
+                    a[k] = bleach.clean(a[k].strip())
+                if a['first_name'] == '':
+                    raise forms.ValidationError("First name must be given for each author.")
+                if a['last_name'] == '':
+                    raise forms.ValidationError("Last name must be given for each author.")
+                if a['orcid_id'] != '' and not re.match('^\d{4}-\d{4}-\d{4}-\d{3}[0-9X]$', a['orcid_id']):
+                    raise forms.ValidationError("ORCID ID must match pattern xxxx-xxxx-xxxx-xxxy, where x is a digit "
+                                                "and y a digit or the capital letter X.")
 
-        cleaned_data['authors'] = bleach.clean(authors_string)  # prevent XSS attacks
+                if len(a['affiliations']) > settings.PUBLICATION_MAX_NUM_AFFILIATIONS_PER_AUTHOR:
+                    raise forms.ValidationError(
+                        f"Too many affiliations given, at maximum {settings.PUBLICATION_MAX_NUM_AFFILIATIONS_PER_AUTHOR}"
+                        "allowed per author.")
+
+                new_affs = []
+                for aff in a['affiliations']:
+                    try:
+                        aff['name'] = bleach.clean(aff['name'].strip())
+                        name_given = len(aff['name']) > 0
+                    except KeyError:
+                        name_given = False
+                    try:
+                        aff['ror_id'] = bleach.clean(aff['ror_id'].strip())
+                        ror_id_given = len(aff['ror_id']) > 0
+                    except KeyError:
+                        ror_id_given = False
+
+                    if name_given:
+                        if ror_id_given and not re.match('^0[^ilouILOU]{6}\d{2}$', aff['ror_id']):
+                            raise forms.ValidationError(
+                                f"Incorrect format for ROR ID \'{aff['ror_id']}\', should start with 0 (zero), followed "
+                                "by 6 characters and should end with 2 digits."
+                            )
+                        new_affs.append(aff)  # only this one should be used, empty affiliations will be ignored
+                    else:
+                        if ror_id_given:
+                            raise forms.ValidationError(
+                                f"Please specify a name for affiliation with ROR ID {aff['ror_id']}."
+                            )
+                a['affiliations'] = new_affs
+
+            #
+            # Brute force search for duplicates
+            #
+            for a_idx, a in enumerate(authors):
+                for b in authors[a_idx+1:]:
+                    if a == b:
+                        raise forms.ValidationError("Duplicate author given! Make sure authors differ "
+                                                    "in at least one field.")
+        except forms.ValidationError:
+            raise
+        except Exception as exc:
+            msg = f"Problems while parsing authors' data: {exc}"
+            _log.error(msg)
+            raise forms.ValidationError("Problems while parsing authors' data.")  # we provide no more detail here
+
+        cleaned_data['authors_json'] = authors
 
         return cleaned_data